requirements.txt
github3.py==4.0.1
numpy==2.2.4
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
Flask==3.0.3
Flask-Cors==4.0.0
Werkzeug==3.0.3
gunicorn==22.0.0

